Retrieve Test Information From DB
This is a quite rudimentary approach to moving all of the test information in the database into a class.

Information can be accessed through the single testArray array as it has class object attributes. The hierarchy is Test->Questions->Answers

This can be seen in test line 61, retrieving the third answer of the first question of the first test.