Basics of Test Creation
Tests can now be created using form fields.

This is not a dynamic form, this is on the list to be implemented, allowing the user to add as many questions as they want on the form when making a test.

Currently it supports a fixed amount of questions and answers per Test.

Structure of code is messy, this is also on the to do list.