Gap issue has been resolved
Fixed the issue with the white space in-between the circles and the second section.
I fixed this by adding the class "indexCircles" to the circles so I can change the height specifically for the index page without affecting other pages.
I also increased the speed slightly for the animation so the circles reach the top of the page on Index. This does affect other pages but is a very slight change which does not result in repeated code.