Updated the first section
The styling for the first section is mostly done, although I might add an image or some other content to help fill space and make the page look more enticing.
Most of the code here is to do with the scroll down animation. When you enter the page some text and an arrow will float up from the bottom of the page.
The content is not finished and I am still using placeholder text.