Changed background image and fixed issues
I changed the background image, for now this is still temporary but it suits the page better than the previous image and it would be easier to work with while building the style of the page.
Also in the previous commit the content in indexScrollDown was not aligned correctly. Now the arrow appears correctly in the middle.