Initial Login Page Commit
I have removed the previous login page branch and created this one as the CSS was out of date from the staging branch and there wasn't much progress on the page.
I have created this branch using Fin's Registration page so it will have a similar style. The current login page is just a placeholder as I work on the functionality of the login page.