Added Content To Fill The Features Boxes
I have added content to fill the 4 features boxes that I created in a previous commit. This helps to fill out the page more.
My next steps is to work on the styling for the boxes. They should be the same size, the text could be bigger and I could add images to help showcase the features further.