Modify Form now populates with Correct Test Info.
When clicking the Modify button on a form, it will now populate the fields with the existing information in the database.

When closing the Modal, the form's contents will reset to the default new test state.

The form can have two "data-mode" states, "edit" and "create". These are assigned depending on if the user clicked the new test or modify test button.

These will be used to differentiate the ajax calls to
either the createTest.php file or the new modifyTest.php file.