Restructuring the index page
I imported the code from the index page onto the new template. I want to go through the code I've already done to try and slimline it by removing unnecessary code.
The issue with the gap on the circles came back again so I have changed the position to fixed instead of absolute, this means that it will not be affected by scrolling which will fix all issues to do with the circles background. It also gives the page a parallax effect which looks better.