Initial Login Page Commit
I have remade the login form because the template and registration page was rebuilt.
I used the Registration page as a template so it has a similar style.
My next steps is to work on the back end so you can actually log in and I need to edit the navigation bar so it hides other elements as you aren't supposed to be logged in if you're on this page.