Frontend of Form is now Dynamic
When clicking add question on form, the accordion element for the first question will be cloned, all of the uniquely identifying stuff such as names will be modified to reflect the new question number and the question will be appended to the form.

When the user clicks the delete question button, the individual accordion element will be removed, and all of the accordion elements beyond that question will have their uniquely identifying elements updated, effectively removing 1 from the numbers to reflect the question before it being removed, keeping the question numbers correct no matter which question has been deleted.