When modifying existing test, deleting a question removes it from DB
When modifying an existing test, deleting a question will delete it from the database.

It will also decrement all relativeQuestionIDs after that point by one to keep questions in order.

If using the form to create a new question, it won't touch the database as the question doesn't even exist yet.