Created Functionality for the Login Page
Here is the base functionality for the login page. Currently it is the bare minimum as I want to make it redirect to the correct pages once the user has logged in. I would also like to add pop ups which show if the user has not logged in successfully, so this code will be changed further.
One last thing to note, the auth level when you've logged in is temporary for now as it will likely be tied to accessLevel in the database for admin purposes.