Modification of Tests now Functional
Have reverted the last change, all modifications remain client side until they press submit on the form.

This will post all data to modifyTest.php

This deletes the test, and all of its questions and answers, rebuilds it with updated data using createTest.php as an include, and then sets the new test's testID to what it was before.

This is hardly an ideal way of doing this but we are under time pressure and this is the most efficient way of this working considering that.