Authorization Added and Sweet Alerts Fixed
I have added authorisation now so when the user logs in, some of their information will be stored to use on other pages. This can be used to restrict access to pages like the admin pages. If the user does not have the correct authorisation level then they will be redirected to another page.
Lastly, the sweet alerts have been updated so they now show the correct information on errors, there is a redirect on the login alert which will take you to the index page, and the username/password error message is changed to be a little bit more ambiguous now in an attempt to prevent malicious users guessing account names and passwords.