Setting up the question style
I am setting up the container for the testing page and have added the question div along with the styling for it. The question div uses similar styling to the features div on the index page, it will be contained in a box similar to how the Registration and Login page looks.