Progress towards passing the testID to testing.php
I am currently having issues with retrieving the testID from test-selection.php to testing.php so I wanted to commit what I've currently got so other project members can assist.
I have set up the jRoute for testing.php, and it did work on a previous test.
I have modified Fin's outputStudentTests.php so when the button is pressed, it can pass along the testID. It currently shows up as a hidden value on the card.
The AJAX code is to detect when the button on testForm is clicked, it will use the jRoute to redirect to testing.php, pass along the data and then output the testID in a console log.
Currently there seems to be a 404 error that appears when the button is clicked.