Finished setting up the testID variable + extra security
I have changed the javascript to link to a php file which sets the test ID in a session variable, this is so the variable can be accessed across the pages. This is better than using the url to get the ID because the user would have been able to change the ID themselves which could lead to complications.
Next steps is to output the data in each box.