jRoute Support for Multiple Roles
jRoute will now check to see if it has been passed an array when the Route method is called, allowing a route to have multiple roles accessing it (as highlight in issue #48)

A new class has been made "User". This stores information about the user to the session with attributes such as their access level, their email and more.

jRoute, when passed an array, will check the accessLevel attribute of the user object stored in the session against the array of allowed roles as defined in the route file (index.php). If the user is one of the roles that are allowed, they will authorize and be redirected, if they aren't, they will receive a 403.