Added relative IDs in the test data
For the testing page, we need relative Question ID and relative Answer ID to correctly display the questions and answers. This data was not being stored in TestArray so I have now added them.
relativeAnswerID is used to order from answers for each question, the IDs will be ordered from 0-3, as an example, an answer with the ID of 2 will be the 3rd answer in the list.
relativeQuestionID is similar to this but it is used to order the questions, so a 10 question exam will have the IDs from 0-9, and the 8th question for example will have the ID of 7.