Setting up the testing framework
In this commit, I have created initialiseQuestions.php. When testing.php is loaded, some jQuery will detect the page and send a POST request to initialiseQuestions.php.
This new php file will access test information within retrieveTests.php and use that to output the test information.
Currently there is only temporary code which was used to test this method of outputting the test. I plan to create a frontend similar to test-selection in the way that this code will output the form in an echo.
After this, there will be another file which will be used to check the user's progress in the test and output the correct question based on their progression.