Reorganised Testing.php and Created Basic Validation
The html code within initialiseQuestions.php has been slimmed down to remove unnecessary code (like the form tag and the temporary submit button). I also made additions to it so the user can now see the current question they are on as well as the subject the exam is under.
validateAnswer is currently pretty basic, currently, but from here I can work on transferring the user to the next question in the exam.