Created SweetAlerts and Fixed Critical Bug
I have created Sweet Alerts which will inform the user whether they have correctly answered the question or not. If the user has not correctly answered the question, the Sweet Alert will include the correct answer so they can learn from their mistakes.
While building this, I discovered a critical issue to do with the jQuery. Since the jQuery code was looking for whenever someone clicked a button with a specific class, if there was another button with the same class on a different page, it would call that function. For example, when I clicked submit on the login page, a Sweet Alert appeared with an error message like I answered the question incorrectly.
To remedy this issue, I have assigned an ID to each button which is called upon in the jQuery code, so functions will not overlap.