Added Functionality to Proceed to Next Question
I have created a new php file which will allow the user to proceed to the next question. This file looks similar to initialiseQuestions.php but it doesn't have the lines at the top which define some $_SESSION variables. I may change how the question display system works later on though since 99% of the code in this file is repeated code.
The second major change I've made within this commit is to do with how the questions are loaded. Previously the code would just replace what was in the <div> tag, but this meant that during the time that it takes to load the question, the user could click on a button again for the same question, this could cause complications like cheating the score. The resolution to this issue was to clear the <div> tag when the Sweet Alert appears. This means that while the user is reading the information the question will disappear in preparation for the next question.
Lastly, I have made the testing interface wider as it only took up half of the page, this could break the page if some of the answers were too long, so this change will make the UI cleaner.