Reworking the Backend
After the previous commit, I decided to rework how outputting the next question is handled since it didn't take too much effort to do. I set the $_SESSION variables in what used to be assignTestID.php (now initialiseQuestions.php) so that ended up making initialiseQuestions.php the exact same as nextQuestion.php, so I removed one of them and editted the script file to reflect this.