Fixing Sweet Alerts Issue
There was an issue on the testing page where you could click on the outside of a Sweet Alert and it will close the Alert without continuing the code. This resulted in a softlock for the testing page. I have added allowOutsideClick: false, so now when the user clicks outside the alert, it will not close. This problem may exist on other pages so I will create an issue on the project page.