Added Checks for Question Number
Once the user has gone through each question in the program, the program will alert the user through a Sweet Alert telling them that they have finished and it will also output the overall score.
For now, it redirects to the test-selection page. I intend on adding one more php file which will finalise the test so the data is updated in the database.