Backend built for retrieving leaderboard information.
retrieveLeaderboardData will now return a class object of LeaderboardData ready to be rendered on the backend.

It works using multidimensional arrays nested inside the class properties.

For example, you could access leaderboardData->computing which would return a multidimensional array with each index having values of userID and Score.

Score is calculated based on the user's total score for that subject.