Finishing touches to merge to staging
In this commit I've updated my code so it works with the code from staging.
This includes the routing changes along with a change to test-selection to link it to testing.php.
Sam did request that I upload the subjectID when finalising the test so the program now stores the SubjectID in a _SESSION variable, uploads it in the sql statement and then unsets the variable when it is no longer required.