Final fix for the scroll down issue
I have put the scroll down text in it's own <div> tag which separates it from the main container, this allows the page to center the content within that container and when it scales the size, it won't affect the width of the container anymore. I have also removed some unnecessary code from the CSS to clean it up.