Sanitizing question/answer text
There was a major issue with the questions and answers where they were not being outputted in plaintext, this could be a critical security issue since any user is able to create their own exams. I have added extra lines which converts the data into plaintext so they will be outputted correctly.