Questions now display in a random order
Questions within the testing screen are now displayed in a random order.

We're generating an array based on the number of questions, so for 5 questions it'll make [0,1,2,3,4] then shuffle it to get a random order.

If we are showing the user their third question for example, we will access the random array at index [2], which will give us a random question.