Unsetting the random order session variable on test init
When quitting a test halfway through, this array would persist in the session. This would cause the next test to try and access array indexes of $test->question[] that were outside the available range.

On test init, we are now unsetting this session var so it can be recreated properly in nextQuestion.php